requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
PyYAML==6.0.2